Associate comment approve and comment delete function to respective user.

+++ blog/views.py
@@ -21,7 +21,7 @@ def register(request):
     return render(request, 'registration/register.html', {'form': form})
 
 def post_list(request):
-    posts = Post.objects.filter(published_date__lte=timezone.now()).order_by('published_date')
+    posts = Post.objects.filter(published_date__lte=timezone.now()).order_by('-published_date')
     return render(request, 'blog/post_list.html', {'posts': posts})
 
 def post_detail(request, pk):
@@ -90,11 +90,13 @@ def add_comment_to_post(request, pk):
 @login_required
 def comment_approve(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    comment.approve()
+    if comment.post.author == request.user:
+        comment.approve()
     return redirect('post_detail', pk=comment.post.pk)
 
 @login_required
 def comment_remove(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    comment.delete()
+    if comment.post.author == request.user:
+        comment.delete()
     return redirect('post_detail', pk=comment.post.pk)
